1. Changed parameter input to wy_file_append() to unsigned int from unsigned short for optimization purpose.
2. Added min and max size checks to the internal buffer.

+++ src/WY_File_IO.c
@@ -2,7 +2,7 @@
 #include "WY_File_IO.h"
 
 
-void wy_file_append(const unsigned short p_size, const char * restrict p_content, FILE *p_fp)
+void wy_file_append(const unsigned int p_size, const char * restrict p_content, FILE *p_fp)
 {
     fwrite(p_content, 1, p_size, p_fp);
 }
